+++ src/repl.ts
@@ -79,7 +79,6 @@ export async function startTestBotRepl(testBot: TestBot): Promise<void> {
 
   try {
     while (true) {
-      // TODO: handle Ctrl+D
       const text = await rl.question("> ");
       if (!text || text === "/quit") {
         break;
